Changed text_processing to text_processor for better compatibility
